+++ browser-switch/src/main/java/com/braintreepayments/api/BrowserSwitchClient.java
@@ -8,6 +8,8 @@ import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
 import androidx.fragment.app.FragmentActivity;
 
+import com.braintreepayments.api.browserswitch.R;
+
 import org.json.JSONObject;
 
 /**
@@ -38,7 +40,7 @@ public class BrowserSwitchClient {
      * Open a browser or <a href="https://developer.chrome.com/multidevice/android/customtabs">Chrome Custom Tab</a>
      * with a given set of {@link BrowserSwitchOptions} from an Android activity.
      *
-     * @param activity the activity used to start browser switch
+     * @param activity             the activity used to start browser switch
      * @param browserSwitchOptions {@link BrowserSwitchOptions} the options used to configure the browser switch
      */
     public void start(@NonNull FragmentActivity activity, @NonNull BrowserSwitchOptions browserSwitchOptions) throws BrowserSwitchException {
@@ -56,7 +58,8 @@ public class BrowserSwitchClient {
         persistentStore.putActiveRequest(request, appContext);
 
         if (browserSwitchInspector.deviceHasChromeCustomTabs(appContext)) {
-            customTabsInternalClient.launchUrl(activity, browserSwitchUrl);
+            boolean launchAsNewTask = browserSwitchOptions.isLaunchAsNewTask();
+            customTabsInternalClient.launchUrl(activity, browserSwitchUrl, launchAsNewTask);
         } else {
             Intent launchUrlInBrowser = new Intent(Intent.ACTION_VIEW, browserSwitchUrl);
             activity.startActivity(launchUrlInBrowser);
@@ -73,22 +76,14 @@ public class BrowserSwitchClient {
         String errorMessage = null;
 
         if (!isValidRequestCode(requestCode)) {
-            errorMessage = "Request code cannot be Integer.MIN_VALUE";
+            errorMessage = activity.getString(R.string.error_request_code_invalid);
         } else if (returnUrlScheme == null) {
-            errorMessage = "A returnUrlScheme is required.";
+            errorMessage = activity.getString(R.string.error_return_url_required);
         } else if (!browserSwitchInspector.isDeviceConfiguredForDeepLinking(appContext, returnUrlScheme)) {
-            errorMessage =
-                "The return url scheme was not set up, incorrectly set up, " +
-                "or more than one Activity on this device defines the same url " +
-                "scheme in it's Android Manifest. See " +
-                "https://github.com/braintree/browser-switch-android for more " +
-                "information on setting up a return url scheme.";
+            errorMessage = activity.getString(R.string.error_device_not_configured_for_deep_link);
         } else if (!browserSwitchInspector.deviceHasBrowser(appContext)) {
-            StringBuilder messageBuilder = new StringBuilder("No installed activities can open this URL");
-            if (browserSwitchUrl != null) {
-                messageBuilder.append(String.format(": %s", browserSwitchUrl.toString()));
-            }
-            errorMessage = messageBuilder.toString();
+            String urlString = (browserSwitchUrl != null) ? browserSwitchUrl.toString() : "";
+            errorMessage = activity.getString(R.string.error_browser_not_found, urlString);
         }
 
         if (errorMessage != null) {
@@ -102,10 +97,10 @@ public class BrowserSwitchClient {
 
     /**
      * Deliver a pending browser switch result to an Android activity.
-     *
+     * <p>
      * We recommend you call this method in onResume to receive a browser switch result once your
      * app has re-entered the foreground.
-     *
+     * <p>
      * Cancel and Success results will be delivered only once. If there are no pending
      * browser switch results, this method does nothing.
      *
@@ -138,10 +133,10 @@ public class BrowserSwitchClient {
 
     /**
      * Peek at a pending browser switch result to an Android activity.
-     *
+     * <p>
      * We recommend you call this method in onResume to receive a browser switch result once your
      * app has re-entered the foreground.
-     *
+     * <p>
      * This can be used in place of {@link BrowserSwitchClient#deliverResult(FragmentActivity)} when
      * you want to know the contents of a pending browser switch result before it is delivered.
      *
@@ -167,5 +162,48 @@ public class BrowserSwitchClient {
         }
 
         return result;
+    }
+
+    /**
+     * Deliver a pending browser switch result that was previously captured by another Android activity.
+     * <p>
+     * Success results will be delivered only once. If there are no pending
+     * browser switch results in the cache, this method does nothing.
+     *
+     * @param context the context used to access the cache
+     * @return {@link BrowserSwitchResult}
+     */
+    public BrowserSwitchResult deliverResultFromCache(@NonNull Context context) {
+        Context applicationContext = context.getApplicationContext();
+        BrowserSwitchResult result = persistentStore.getActiveResult(applicationContext);
+        if (result != null) {
+            persistentStore.removeAll(applicationContext);
+        }
+        return result;
+    }
+
+    /**
+     * Capture a pending browser switch result for an Android activity into a persistent storage cache.
+     * <p>
+     * To obtain the result in a separate activity, call {@link #deliverResultFromCache(Context)}.
+     *
+     * @param activity the activity that received the deep link back into the app
+     */
+    public void captureResult(@NonNull FragmentActivity activity) {
+        Intent intent = activity.getIntent();
+        Context appContext = activity.getApplicationContext();
+
+        BrowserSwitchRequest request = persistentStore.getActiveRequest(appContext);
+        if (request == null || intent == null) {
+            // no pending browser switch request found
+            return;
+        }
+
+        Uri deepLinkUrl = intent.getData();
+        if (deepLinkUrl != null) {
+            BrowserSwitchResult result =
+                new BrowserSwitchResult(BrowserSwitchStatus.SUCCESS, request, deepLinkUrl);
+            persistentStore.putActiveResult(result, activity.getApplicationContext());
+        }
     }
 }