+++ browser-switch/src/main/java/com/braintreepayments/api/BrowserSwitchPersistentStore.java
@@ -16,6 +16,9 @@ class BrowserSwitchPersistentStore {
     @VisibleForTesting
     static final String REQUEST_KEY = "browserSwitch.request";
 
+    @VisibleForTesting
+    static final String RESULT_KEY = "browserSwitch.result";
+
     private static final BrowserSwitchPersistentStore INSTANCE = new BrowserSwitchPersistentStore();
 
     static BrowserSwitchPersistentStore getInstance() {
@@ -32,6 +35,7 @@ class BrowserSwitchPersistentStore {
             try {
                 request = BrowserSwitchRequest.fromJson(activeRequestJson);
             } catch (JSONException e) {
+                // NEXT_MAJOR_VERSION: Add explicit error handling instead of ignoring exception
                 Log.d(TAG, e.getMessage());
                 Log.d(TAG, Arrays.toString(e.getStackTrace()));
             }
@@ -43,12 +47,44 @@ class BrowserSwitchPersistentStore {
         try {
             PersistentStore.put(REQUEST_KEY, request.toJson(), context);
         } catch (JSONException e) {
+            // NEXT_MAJOR_VERSION: Add explicit error handling instead of ignoring exception
             Log.d(TAG, e.getMessage());
             Log.d(TAG, Arrays.toString(e.getStackTrace()));
         }
     }
 
+    void putActiveResult(BrowserSwitchResult result, Context context) {
+        try {
+            PersistentStore.put(RESULT_KEY, result.toJson(), context);
+        } catch (JSONException e) {
+            // NEXT_MAJOR_VERSION: Add explicit error handling instead of ignoring exception
+            Log.d(TAG, e.getMessage());
+            Log.d(TAG, Arrays.toString(e.getStackTrace()));
+        }
+    }
+
+    BrowserSwitchResult getActiveResult(Context context) {
+        BrowserSwitchResult request = null;
+
+        String activeResultJSON = PersistentStore.get(RESULT_KEY, context);
+        if (activeResultJSON != null) {
+            try {
+                request = BrowserSwitchResult.fromJson(activeResultJSON);
+            } catch (JSONException e) {
+                // NEXT_MAJOR_VERSION: Add explicit error handling instead of ignoring exception
+                Log.d(TAG, e.getMessage());
+                Log.d(TAG, Arrays.toString(e.getStackTrace()));
+            }
+        }
+        return request;
+    }
+
     void clearActiveRequest(Context context) {
+        PersistentStore.remove(REQUEST_KEY, context);
+    }
+
+    void removeAll(Context context) {
+        PersistentStore.remove(RESULT_KEY, context);
         PersistentStore.remove(REQUEST_KEY, context);
     }
 }